% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAnalyzeVignettes-package.R
\docType{package}
\name{OlinkAnalyzeVignettes-package}
\alias{OlinkAnalyzeVignettes}
\alias{OlinkAnalyzeVignettes-package}
\title{OlinkAnalyzeVignettes: Vignettes that Facilitate Analysis of Proteomic Data from Olink}
\description{
A collection of vignettes to exemplify analysis of proteomic data from Olink, primarily NPX data that has been exported from Olink Software. The functions also work on QUANT data from Olink by log- transforming the QUANT data. The vignettes are focused on reading data, demonstrating data wrangling and quality control analysis, performing statistical analysis and generating figures to visualize the results of the statistical analysis. The goal of this package is to guide users extract biological insights from proteomic data run on the Olink platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://olink.com/} \url{https://github.com/Olink-Proteomics/OlinkRPackage}
}

}
\author{
\strong{Maintainer}: Kathleen Nevola \email{biostattools@olink.com} (\href{https://orcid.org/0000-0002-5183-6444}{ORCID}) (kathy-nevola)

Authors:
\itemize{
  \item Kathleen Nevola \email{biostattools@olink.com} (\href{https://orcid.org/0000-0002-5183-6444}{ORCID}) (kathy-nevola)
  \item Marianne Sandin (\href{https://orcid.org/0000-0001-6186-963X}{ORCID}) (marisand)
  \item Jamey Guess (\href{https://orcid.org/0000-0002-4017-0923}{ORCID}) (jrguess)
  \item Simon Forsberg (\href{https://orcid.org/0000-0002-7451-9222}{ORCID}) (simfor)
  \item Christoffer Cambronero (Orbmac)
  \item Pascal Pucholt (\href{https://orcid.org/0000-0003-3342-1373}{ORCID}) (AskPascal)
  \item Boxi Zhang (\href{https://orcid.org/0000-0001-7758-6204}{ORCID}) (boxizhang)
  \item Masoumeh Sheikhi (MasoumehSheikhi)
  \item Klev Diamanti (\href{https://orcid.org/0000-0002-4922-8415}{ORCID}) (klevdiamanti)
  \item Amrita Kar (amrita-kar)
  \item Lei Conze (leiliuC)
  \item Kristyn Chin (kristynchin-olink)
  \item Danai Topouza (\href{https://orcid.org/0000-0002-6897-9281}{ORCID}) (dtopouza)
  \item Stephen Pollo (\href{https://orcid.org/0000-0001-9252-4976}{ORCID}) (spollo-olprot)
  \item Kang Dong (\href{https://orcid.org/0000-0002-4567-5007}{ORCID}) (KangD-dev)
}

Other contributors:
\itemize{
  \item Kristian Hodén (\href{https://orcid.org/0000-0003-0354-0662}{ORCID}) (kristianHoden) [contributor]
  \item Per Eriksson (\href{https://orcid.org/0000-0001-7633-403X}{ORCID}) (b_watcher) [contributor]
  \item Nicola Moloney (\href{https://orcid.org/0000-0003-4967-3284}{ORCID}) [contributor]
  \item Britta Lötstedt (\href{https://orcid.org/0000-0003-3545-5489}{ORCID}) [contributor]
  \item Emmett Sprecher (\href{https://orcid.org/0000-0002-7710-695X}{ORCID}) [contributor]
  \item Jessica Barbagallo (jbarbagallo) [contributor]
  \item Olof Mansson (olofmansson) [contractor]
  \item Ola Caster (OlaCaster) [contributor]
  \item Olink [copyright holder, funder]
}

}
\keyword{internal}
